Add RangeDB interfaces & comments
